% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_vector_comb.R
\name{all_vector_comb}
\alias{all_vector_comb}
\title{All possible combinations of a vector}
\usage{
all_vector_comb(data)
}
\arguments{
\item{data}{A vector for which you want to find all the possible combinations}
}
\value{
Vector with all the possible combinations of the input vector
}
\description{
Returns all possible combinations of a vector
}
\examples{
all_vector_comb(data = c("a","b","c"))

}
