% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allVectorComb.R
\name{allVectorComb}
\alias{allVectorComb}
\title{All possible combinations of a vector}
\usage{
allVectorComb(data)
}
\arguments{
\item{data}{A vector for which you want to find all the possible combinations}
}
\value{
Vector with all the possible combinations of the input vector
}
\description{
Returns all possible combinations of a vector
}
\examples{
allVectorComb(data = c("a","b","c"))

}
