% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OdbcConnect.SQLServer.R
\name{OdbcConnect.SQLServer}
\alias{OdbcConnect.SQLServer}
\title{Establishes a connection handle to the SQL Server}
\usage{
OdbcConnect.SQLServer(username)
}
\arguments{
\item{username}{Username/RACF to connect to the server}
}
\value{
connection string to the SQL Server to be used to query tables and views
}
\description{
Returns a connection string to the SQL Server to be used to query tables and views
}
