% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rep_seq.R
\name{plot_rep_seq}
\alias{plot_rep_seq}
\title{Sequence plot}
\usage{
plot_rep_seq(seq, x_title = "", y_title = "", title = "",
  text_width = 2)
}
\arguments{
\item{seq}{A dataframe or datatable with each row representing a sequence}

\item{x_title}{A character vector to be displayed as the title of x axis}

\item{y_title}{A character vector to be displayed as the title of y axis}

\item{title}{A character vector to be displayed as the title of the plot}

\item{text_width}{An integer denoting the size of the sequence code}
}
\value{
Plot visualizing the sequences along with the sequence code
}
\description{
Returns a plot visualizing sequences along with the sequence code
}
\examples{
plot_rep_seq(seq = data.frame(1:10,2:11,3:12,4:13,5:14),x_title = "X Axis",y_title = "Y Axis",title = "Plot Title",text_width = 5)

}
