% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeq.R
\name{plotSeq}
\alias{plotSeq}
\title{Sequence plot}
\usage{
plotSeq(seq, x.title = "", y.title = "", title = "",
  text.width = 2)
}
\arguments{
\item{seq}{A dataframe or datatable with each row representing a sequence}

\item{x.title}{A character vector to be displayed as the title of x axis}

\item{y.title}{A character vector to be displayed as the title of y axis}

\item{title}{A character vector to be displayed as the title of the plot}

\item{text.width}{An integer denoting the size of the sequence code}
}
\value{
Plot visualizing the sequences along with the sequence code
}
\description{
Returns a plot visualizing sequences along with the sequence code
}
\examples{
plotSeq(seq = data.frame(1:10,2:11,3:12,4:13,5:14),x.title = "X Axis",y.title = "Y Axis",title = "Plot Title",text.width = 5)

}
