% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSQL.R
\name{readSQL}
\alias{readSQL}
\title{Reads sql queries from a .sql file}
\usage{
readSQL(filepath)
}
\arguments{
\item{filepath}{Full path of the sql file}
}
\value{
A vector of SQL queries present in the .sql file separated by semi colon
}
\description{
Returns A vector of SQL queries present in the .sql file separated by semi colon
}
