% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSQL.R
\name{runSQL}
\alias{runSQL}
\title{Runs all the sql queries for a connection instance}
\usage{
runSQL(con, sql.vector)
}
\arguments{
\item{con}{connection handle as returned by odbcConnect}

\item{sql.vector}{vector of sql queries}
}
\value{
a list of results generated after running the sql queries
}
\description{
Returns a list of results generated after running the sql queries
}
